+++ hassio-addon/CHANGELOG.md
@@ -1,6 +1,8 @@
 ** ⚠️ 업데이트 전 백업 권장 **
 - 오류가 있으면 homeassistant [카페](https://cafe.naver.com/koreassistant), [깃헙](https://github.com/wooooooooooook/homenet2mqtt), [디스코드](https://discord.gg/kGwhUBMe5z) 등으로 알려주세요.
 
+- fix: 시작 직후 UI가 업데이트되지 않는 현상 수정
+
 v3.0.1 - v3.0.4 Hotfix
 - fix: mqtt애드온에 불필요한 포트매핑제거
   - 업데이트후 mqtt애드온이 실행안되시는분은 구성에서 네트워크설정을 기본값으로 재설정해주세요.